rtest: Cast fail+error to int in RTEST_MAIN exit code
main() returns int but RTestReport->fail / ->error are rsize, so
'ret = report->fail + report->error' produced C4244 wherever
RTEST_MAIN was expanded (test/rtest, example/rtest, bench/main).
Add the cast inside the macro so every consumer is silenced
in one place.

+++ include/rlib/rtest.h
@@ -231,7 +231,7 @@ int main (int argc, rchar ** argv) {                                            
         (report = r_test_run_tests (tests, count, run_flags, f,                 \
             R_TEST_ALL_MASK, filter)) != NULL) {                                \
       r_test_report_print (report, report_flags, f);                            \
-      ret = report->fail + report->error;                                       \
+      ret = (int) (report->fail + report->error);                              \
       r_test_report_free (report);                                              \
     }                                                                           \
     if (f != stdout)                                                            \
